server/requirements.txt
aniso8601==9.0.1
astroid==2.12.13
click==8.1.3
colorama==0.4.6
dill==0.3.6
Flask==2.2.2
Flask-API==3.0.post1
Flask-HTTPAuth==4.7.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==5.1.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
platformdirs==2.5.4
pycryptodome==3.16.0
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.44
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
